move get questions handler
